requirements.txt
openai==1.59.9
pandas==2.2.3
pdf2image==1.17.0
Pillow==11.1.0
pytesseract==0.3.13
